Material schemas conflict solved; refreshing working; material querying left
